// src/logistics-api/model/createTruck.ts
/**
 * Logistics API
 * Logistics API provides services to find and manage entities for logistics
 *
 * OpenAPI spec version: 1.0
 * Contact: vladislavprokopenko1@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Object representing Truck creating
 */
export interface CreateTruck { 
    number?: string;
    model?: string;
    /**
     * Truck capacity, tons
     */
    capacity?: number;
    /**
     * Total operating time, years
     */
    totalOperatingTime?: number;
    currentCity?: string;
    currentState?: string;
}